registering multiple instances of user service

+++ routes/user.js
@@ -8,20 +8,39 @@ const registry = require("./registry.json");
 router.post("/register", async (req, res) => {
   try {
     const registryInfo = req.body;
-    registry.services[registryInfo.apiName] = { ...registryInfo };
-    fs.writeFile(
-      "./routes/registry.json",
-      JSON.stringify(registry),
-      (error) => {
-        if (error)
-          res.send(
-            "Could not register '" + registryInfo.apiName + "'\n" + error
-          );
-        else {
-          res.send("Successfully registered '" + registryInfo.apiName + "'");
+    registryInfo.url =
+      registryInfo.protocol +
+      "://" +
+      registryInfo.host +
+      ":" +
+      registryInfo.port +
+      "/users" +
+      "/";
+
+    if (apiAlreadyExists(registryInfo)) {
+      res.send(
+        "Configuration already exists for '" +
+          registryInfo.apiName +
+          "' at '" +
+          registryInfo.url +
+          "'"
+      );
+    } else {
+      registry.services[registryInfo.apiName].push({ ...registryInfo });
+      fs.writeFile(
+        "./routes/registry.json",
+        JSON.stringify(registry),
+        (error) => {
+          if (error)
+            res.send(
+              "Could not register '" + registryInfo.apiName + "'\n" + error
+            );
+          else {
+            res.send("Successfully registered '" + registryInfo.apiName + "'");
+          }
         }
-      }
-    );
+      );
+    }
   } catch (error) {
     console.log(error);
   }
@@ -41,5 +60,13 @@ router.all("/:type", async (req, res) => {
     console.log(error);
   }
 });
+
+const apiAlreadyExists = (registryInfo) => {
+  let exist = false;
+  registry.services[registryInfo.apiName].forEach((instance) => {
+    if (instance.url === registryInfo.url) exist = true;
+  });
+  return exist;
+};
 
 module.exports = router;
